Use a long uid for artery and int for old remoting #20644

+++ akka-remote/src/main/scala/akka/remote/RemoteActorRefProvider.scala
@@ -455,7 +455,7 @@ private[akka] class RemoteActorRefProvider(
    * @param uid UID of the remote system, if the uid is not defined it will not be a strong quarantine but
    *   the current endpoint writer will be stopped (dropping system messages) and the address will be gated
    */
-  def quarantine(address: Address, uid: Option[Int], reason: String): Unit =
+  def quarantine(address: Address, uid: Option[Long], reason: String): Unit =
     transport.quarantine(address, uid, reason)
 
 }
